Bound the metadata length, close two silent-skip gaps, type RecordHeader.RType
Six review fixes, the last work before the branch merges.

1. Bound the declared metadata length. DbnFsm handed MetadataDecoder.DecodePrelude's
   unbounded `length` straight to AlignedBuffer.Grow. Measured against the pre-fix code,
   a declared length up to 2,147,483,632 silently allocated whatever it asked for (2 GiB
   at the top of the band), 2,147,483,633..639 threw OverflowException as the round-up to
   a ulong count wrapped, and 2,147,483,640 and up threw ArgumentOutOfRangeException as
   `length + 8` wrapped negative. None of them a DbnDecodeException, which is the only
   exception DbnDecoder and DbnFsm document for malformed DBN, and this is the first eight
   bytes of an unauthenticated live stream.

   Adds DbnConstants.MaxMetadataLength = 512 MiB, rejected past that with a
   DbnDecodeException naming the length and the limit. Not 16 MiB: an ALL_SYMBOLS
   definition query legitimately produces tens to hundreds of MB. Not higher: 2 GiB is
   where int arithmetic fails and where a forged length is worth mounting as a DoS.
   ToUlongLength and the `length + prelude` addition are now 64-bit so neither can wrap.
   The incremental read that removes the need to trust this field at all stays as #12.

2. Guard the record-size list itself. AssertWireSize folds in the header-first assertion
   and claimed a record "cannot skip this check" — but it is only ever reached from a
   hand-written list of 21 calls, so a new record struct escaped the size, alignment and
   header-first checks alike. The list is now one method run by two tests, the second of
   which reflects over every IRecord<> implementer in the assembly and fails, by name, if
   the two sets differ. Comment corrected to describe what is actually true.

3. Completeness tests for EnumValues and WireStrings. Both are hand-written switches
   paralleled by hand-written [InlineData] rows, complete today and defended by nothing.
   Adds Enum.GetValues<T>() tests over all 20 wire-validated enums and all 7 with a wire
   string, following PublisherTableTests' pattern: every declared variant accepted, and —
   by sweeping the whole 8- or 16-bit domain — nothing else. The seven char-valued enums
   have no wire strings by design and are deliberately absent.

4. Exercise the metadata variable section's multi-element paths. Every non-fragment
   fixture tops out at one symbol, zero partial, zero not_found, one mapping, one
   interval, so every loop in ReadRepeatedCstr/ReadMappings/WriteRepeatedCstr/
   WriteMappings ran at most once in both directions. Adds a hand-built v3 block with
   three symbols, one partial, two not_found, two mappings and two intervals — distinct
   lengths and contents, so a Partial/NotFound transposition fails. Verified: that
   mutation fails all three new tests and none of the 59 existing ones. Also narrows the
   false byte-identity claim (every reserved run in the corpus is zero, so byte-identity
   cannot discriminate a dropped one), narrows EncodedLength's exception doc, and adds an
   encode test into a 0xAA-prefilled destination.

5. RecordHeader joins the RawX/X convention: the field is RawRType and RType is the typed
   view, matching every other record. RTypeSchemaMapping.TryIntoSchema gains a typed
   overload so it has the same shape as its ToRType counterpart, and RecordRef.Has<T> no
   longer casts privately. Free today with no external consumers; a breaking change the
   day M2 compiles against it.

6. CI fails loudly if no 11.x SDK resolves. HasNet11Sdk is inferred from NETCoreSdkVersion
   and falls back to net10.0 alone, so a setup-dotnet miss would take the whole matrix
   green having built nothing on net11.0 — indistinguishable from no signal, on a workflow
   that has never run. Now asserts both an 11.x SDK and net11.0 in the resolved
   TargetFrameworks.

732 -> 788 tests, zero warnings, Debug and Release.

Refs #3 #4 #5

+++ src/DatabentoDotNet.Dbn/Decoding/AlignedBuffer.cs
@@ -213,5 +213,11 @@ public sealed class AlignedBuffer
 
     private Span<byte> AsBytes() => MemoryMarshal.AsBytes<ulong>(_memory.AsSpan());
 
-    private static int ToUlongLength(int byteCapacity) => (byteCapacity + 7) / sizeof(ulong);
+    // The round-up is done in 64-bit on purpose. `(byteCapacity + 7)` wraps negative for the top
+    // seven int values, and the negative quotient then reaches `new ulong[...]` as an
+    // OverflowException — a failure mode that has nothing to do with the caller's mistake and
+    // that no `catch (DbnException)` in the decoder would ever see. Widening makes the result
+    // always the true ceiling; a capacity the machine genuinely cannot satisfy then fails as an
+    // ordinary allocation failure instead.
+    private static int ToUlongLength(int byteCapacity) => (int)(((long)byteCapacity + 7) / sizeof(ulong));
 }
